preprocessing/requirements.txt
librosa~=0.7.2
pydub==0.25.1
PyYAML==5.4.1
nltk~=3.5
spacy==3.0.5
gensim==4.0.0
numba~=0.48
python_speech_features